Migrate anime service methods to anime admin service for better admin functionality
